requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
httpx==0.28.1
python-dotenv==1.2.2
